requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.0.1
blinker==1.9.0
cachetools==5.5.2
catboost==1.2.8
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
cloudpickle==2.2.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
databricks-sdk==0.55.0
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
docker==7.1.0
executing==2.2.0
fastapi==0.115.8
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.0
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.20.3
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
importlib_metadata==8.6.1
ipinfo==5.1.1
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lightgbm==4.6.0
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.22.0
mlflow-skinny==2.22.0
multidict==6.4.4
narwhals==1.41.0
numpy==2.2.6
nvidia-nccl-cu12==2.26.5
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
optuna==4.3.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==11.2.1
plotly==6.1.2
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.9.1
pydantic_core==2.33.0
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.7.0
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-rapidjson==1.14
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
secure==0.3.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.18
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.45.3
streamlit==1.45.1
tenacity==9.1.2
threadpoolctl==3.6.0
tldextract==5.1.3
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xgboost==3.0.2
yarl==1.20.0
zenml==0.83.0
zipp==3.22.0
